fix: Suppress all MSVC compiler warnings

+++ src/data_source/imu_capture.cpp
@@ -70,8 +70,8 @@ void imu_capture::start(std::function<void(const pose_frame &)> frame_received) 
             data.gyro_status = frame.imu[i].gyro;
             data.mag_status = frame.imu[i].mag;
             data.orientation =
-                glm::quat(frame.imu[i].orientation_quat.w, frame.imu[i].orientation_quat.x,
-                          frame.imu[i].orientation_quat.y, frame.imu[i].orientation_quat.z);
+                glm::quat(static_cast<float>(frame.imu[i].orientation_quat.w), static_cast<float>(frame.imu[i].orientation_quat.x),
+                          static_cast<float>(frame.imu[i].orientation_quat.y), static_cast<float>(frame.imu[i].orientation_quat.z));
             pose.poses.push_back(data);
           }
